Changed phase in which module (class) is checked
We checked existence of necessary methods in BEGIN phase which leads to
constant fails when modules is imported in another module. It was
mistake: Perl couldn't identify methods defined _after_ BEGIN block
which located almost in the beginning of every module with check. It's
better ro use CHECK phase because it intended to literally check the
code in the module.

Signed-off-by: Georgiy Odisharia <math.kraut.cat@gmail.com>